feat:upgrade go version and solve unreasonable package dependencies
