+++ tools/wasme/cli/example/tinygo/go.mod
@@ -2,6 +2,4 @@ module main
 
 go 1.15
 
-require (
-	github.com/tetratelabs/proxy-wasm-go-sdk v0.1.1
-)
+require github.com/tetratelabs/proxy-wasm-go-sdk v0.1.1
