+++ tools/wasme/pkg/defaults/defaults.go
@@ -2,6 +2,8 @@ package defaults
 
 import (
 	"context"
+	"os"
+	"path/filepath"
 
 	"github.com/solo-io/wasm/tools/wasme/pkg/cache"
 	"github.com/solo-io/wasm/tools/wasme/pkg/pull"
@@ -20,4 +22,15 @@ func NewDefaultPuller() pull.ImagePuller {
 	// Can pull from non-private registries
 	res, _ := resolver.NewResolver("", "", true, false)
 	return pull.NewPuller(res)
+}
+
+var (
+	WasmeConfigDir       = home() + "/.wasme"
+	WasmeImageDir        = filepath.Join(WasmeConfigDir, "store")
+	WasmeCredentialsFile = filepath.Join(WasmeConfigDir, "credentials.json")
+)
+
+func home() string {
+	dir, _ := os.UserHomeDir()
+	return dir
 }
